Refactor Statiq ScripObject and make it use Document Accessor internally.
